requirements.txt
discord.py
requests
python-dotenv
easy-pil
asyncio
